requirements.txt
google-genai
svgwrite
python-dotenv
Pillow
langgraph
langfuse
